Monior fixes to Utilities Library Added Doxyfile to generate documentation Fixed typo on geFlags that prevented the creation of flags from enumarator using the macro Added virtual destructor to RTTIField to avoid memory leaks while deriving from this class Remover empty repeated StringUtil template definitions Fixed array dimensionalization to convert to hex to literals

+++ geUtility/Include/geRTTIType.h
@@ -1673,7 +1673,11 @@ namespace geEngineSDK {
       RTTIReflectablePtrField<DataType, ObjectType>*
       newField = ge_new<RTTIReflectablePtrField<DataType, ObjectType>>();
 
-      newField->initSingle(name, uniqueId, getter, setter, flags);
+      newField->initSingle(name,
+                           static_cast<uint16>(uniqueId),
+                           getter,
+                           setter,
+                           flags);
       addNewField(newField);
     }
